+++ naga/src/front/spv/mod.rs
@@ -177,7 +177,7 @@ bitflags::bitflags! {
 
 impl DecorationFlags {
     fn to_storage_access(self) -> crate::StorageAccess {
-        let mut access = crate::StorageAccess::all();
+        let mut access = crate::StorageAccess::LOAD | crate::StorageAccess::STORE;
         if self.contains(DecorationFlags::NON_READABLE) {
             access &= !crate::StorageAccess::LOAD;
         }
@@ -4490,6 +4490,7 @@ impl<I: Iterator<Item = u32>> Frontend<I> {
                 | S::Store { .. }
                 | S::ImageStore { .. }
                 | S::Atomic { .. }
+                | S::ImageAtomic { .. }
                 | S::RayQuery { .. }
                 | S::SubgroupBallot { .. }
                 | S::SubgroupCollectiveOperation { .. }