+++ naga/src/valid/mod.rs
@@ -152,6 +152,8 @@ bitflags::bitflags! {
         /// [`AtomicFunction::Exchange { compare: None }`]: crate::AtomicFunction::Exchange
         /// [`Storage`]: crate::AddressSpace::Storage
         const SHADER_FLOAT32_ATOMIC = 1 << 21;
+        /// Support for atomic operations on images.
+        const TEXTURE_ATOMIC = 1 << 22;
     }
 }