+++ naga/tests/snapshots.rs
@@ -789,6 +789,10 @@ fn convert_wgsl() {
             Targets::SPIRV | Targets::METAL | Targets::HLSL | Targets::WGSL,
         ),
         (
+            "atomicTexture",
+            Targets::SPIRV | Targets::METAL | Targets::GLSL | Targets::HLSL | Targets::WGSL,
+        ),
+        (
             "atomicOps-float32",
             Targets::SPIRV | Targets::METAL | Targets::WGSL,
         ),