+++ wgpu-core/src/device/mod.rs
@@ -426,6 +426,10 @@ pub fn create_validator(
         features.contains(wgt::Features::SHADER_INT64_ATOMIC_ALL_OPS),
     );
     caps.set(
+        Caps::TEXTURE_ATOMIC,
+        features.contains(wgt::Features::TEXTURE_ATOMIC),
+    );
+    caps.set(
         Caps::SHADER_FLOAT32_ATOMIC,
         features.contains(wgt::Features::SHADER_FLOAT32_ATOMIC),
     );